update qwen3.5 converter
