+++ lmdeploy/turbomind/models/__init__.py
@@ -1,5 +1,4 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from ._qwen3_5 import _Qwen3_5Model  # noqa: F401
 from .glm4_moe_lite import Glm4MoeLiteModel  # noqa: F401
 from .gpt_oss import GptOssModel  # noqa: F401
 from .internlm2 import InternLM2Model  # noqa: F401
@@ -7,4 +6,4 @@ from .internvl import InternVLModel  # noqa: F401
 from .llama import LlamaModel  # noqa: F401
 from .qwen2 import Qwen2Model  # noqa: F401
 from .qwen3 import Qwen3TextModel  # noqa: F401
-from .qwen3_5 import Qwen3_5Model  # noqa: F401
+from .qwen3_5 import Qwen3_5Model, Qwen3_5TextModel, Qwen3_5VisionModel  # noqa: F401
